# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DigiProducts API — single-tenant AI digital product creation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{userId}:
    patch:
      operationId: updateUser
      tags: [admin]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{userId}/credits:
    post:
      operationId: grantCredits
      tags: [admin]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditGrant"
      responses:
        "200":
          description: Granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/invitations:
    get:
      operationId: getInvitations
      tags: [admin]
      responses:
        "200":
          description: All invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
    post:
      operationId: createInvitation
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationInput"
      responses:
        "201":
          description: Invitation created (includes accept URL)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
...
  /admin/access-requests/{requestId}:
    patch:
      operationId: updateAccessRequest
      tags: [admin]
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
...
  /admin/credit-costs:
    get:
      operationId: getCreditCosts
      tags: [admin]
      responses:
        "200":
          description: Cost per operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditCost"
    patch:
      operationId: updateCreditCosts
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditCostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditCost"
...
  /admin/brand-kit:
    patch:
      operationId: updateBrandKit
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandKitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandKit"
  /admin/settings:
    get:
      operationId: getSettings
      tags: [admin]
      responses:
        "200":
          description: Platform settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
    patch:
      operationId: updateSettings
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
  /admin/audit-logs:
    get:
      operationId: getAuditLogs
      tags: [admin]
      parameters:
        - name: action
          in: query
          schema: { type: string }
        - name: actorId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
components:
  schemas:
...
    AccessRequest:
      type: object
      required: [id, name, email, status, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        requestedRole: { type: ["string", "null"] }
        message: { type: ["string", "null"] }
        status: { type: string, enum: [new, invited, dismissed] }
        createdAt: { type: string }
    AccessRequestUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [invited, dismissed] }
...
    CreditGrant:
      type: object
      required: [amount]
      properties:
        amount: { type: integer }
        note: { type: string }
    CreditCost:
      type: object
      required: [operation, cost]
      properties:
        operation: { type: string }
        cost: { type: integer }
        label: { type: ["string", "null"] }
    CreditCostUpdate:
      type: object
      required: [costs]
      properties:
        costs:
          type: array
          items:
            $ref: "#/components/schemas/CreditCost"
...
    User:
      type: object
      required: [id, email, fullName, role, status, creditsBalance, onboardingComplete, onboardingExempt, createdAt]
      properties:
        id: { type: string }
        email: { type: string }
        fullName: { type: string }
        role: { type: string }
        status: { type: string, enum: [invited, active, suspended, deactivated] }
        creditsBalance: { type: integer }
        onboardingComplete: { type: boolean }
        onboardingExempt: { type: boolean }
        lastLoginAt: { type: ["string", "null"] }
        createdAt: { type: string }
    UserUpdate:
      type: object
      properties:
        role: { type: string, enum: [admin, manager, creator, uploader, marketer] }
        status: { type: string, enum: [active, deactivated] }
        onboardingExempt: { type: boolean }
        resetOnboarding: { type: boolean }
    Invitation:
      type: object
      required: [id, email, role, status, createdAt]
      properties:
        id: { type: string }
        email: { type: string }
        role: { type: string }
        status: { type: string, enum: [pending, accepted, revoked, expired] }
        acceptUrl: { type: ["string", "null"] }
        expiresAt: { type: ["string", "null"] }
        invitedByName: { type: ["string", "null"] }
        createdAt: { type: string }
    InvitationInput:
      type: object
      required: [email, role]
      properties:
        email: { type: string }
        role: { type: string, enum: [admin, manager, creator, uploader, marketer] }
...
    BrandKit:
      type: object
      required: [primaryColor, secondaryColor, accentColor, headingFont, bodyFont]
      properties:
        logoUrl: { type: ["string", "null"] }
        primaryColor: { type: string }
        secondaryColor: { type: string }
        accentColor: { type: string }
        headingFont: { type: string }
        bodyFont: { type: string }
        defaultAuthor: { type: ["string", "null"] }
        footerText: { type: ["string", "null"] }
        defaultDisclaimer: { type: ["string", "null"] }
    BrandKitUpdate:
      type: object
      properties:
        logoUrl: { type: string }
        primaryColor: { type: string }
        secondaryColor: { type: string }
        accentColor: { type: string }
        headingFont: { type: string }
        bodyFont: { type: string }
        defaultAuthor: { type: string }
        footerText: { type: string }
        defaultDisclaimer: { type: string }
    PlatformSettings:
      type: object
      required: [sequentialUnlock, allowManualComplete, approvalWorkflowEnabled, uploaderCanGenerate]
      properties:
        sequentialUnlock: { type: boolean }
        allowManualComplete: { type: boolean }
        approvalWorkflowEnabled: { type: boolean }
        uploaderCanGenerate: { type: boolean }
        managerWeeklyGrantCap: { type: integer }
    PlatformSettingsUpdate:
      type: object
      properties:
        sequentialUnlock: { type: boolean }
        allowManualComplete: { type: boolean }
        approvalWorkflowEnabled: { type: boolean }
        uploaderCanGenerate: { type: boolean }
        managerWeeklyGrantCap: { type: integer }
    AuditLog:
      type: object
      required: [id, action, createdAt]
      properties:
        id: { type: string }
        actorName: { type: ["string", "null"] }
        action: { type: string }
        entityType: { type: ["string", "null"] }
        entityId: { type: ["string", "null"] }
        summary: { type: ["string", "null"] }
        ip: { type: ["string", "null"] }
        createdAt: { type: string }